==> Turner_dataset_reanalysis/5_immcantation_reticulate_full_dataset.Rmd <==
---
title: "Immuncantation reticulation - full Turner dataset"
author: "Ed Carr, Linterman lab"
date: "11 July 2022"
output: 
  pdf_document: 
editor_options: 
  chunk_output_type: console
---

# Motivation

This Rmd will download, install and execute IgBLAST (IGMT references) and Immcantation suite.

Downloading the Immcantation docker image would probably be as effective, if you're more familiar with Docker.

Here we need to VDJ assign for the whole of Turner et al., to trak clones from PBMC and FNA.


```{r setup, include=FALSE}

## set a few options for pdf formatting.
knitr::opts_chunk$set(echo = F, tidy = T, warning = F, message = F, fig.width = 7, fig.height = 9.5, tidy.opts=list(width.cutoff=60))
```

```{r}
# Use conda + bioconda to install IgBLAST
# Note that v1.15 is specified
# Advantage here is that IgBLAST is preconfigured > v1.13 (it is tedious to do by hand)

###
# set the name of the conda environment
envname <- "igblast_changeo_env"
# rest of chunk should run itself
###

library(reticulate)

if (envname %in% conda_list()$name) {
  message("Conda environment already exists. Skipping.\nIf a fresh environment desired, delete the current one or specify a new name")
  conda_env_finalise <- FALSE
  
  } else {

    conda_env_finalise <- TRUE
conda_create(envname = "igblast_changeo_env") # this will over-write any conda environment of same name.

conda_install(envname = "igblast_changeo_env",
              packages = c("igblast=1.15", "changeo=1.0.1", "igphyml=1.1.3", "phylip=3.697"),
              channel = "bioconda")
}
use_condaenv("igblast_changeo_env", required = TRUE)


## It is useful to have the path to this environment as an R variable.
conda_path <- conda_list()

conda_path <- conda_path[conda_path$name == "igblast_changeo_env", "python"]
conda_path <- gsub(conda_path, pattern = "/bin/python", replacement = "")

## The next issue is that system() does not 'see' this conda environment
## system() runs in whichever shell terminal R or Rstudio creates
## whereas the reticulate::python os module does see the right path.

```

```{python, eval = F}

# This is an option to run python code in rstudio.
import os
os.system("which AssignGenes.py")  # AssignGenes.py is in changeo suite.
exit

```

```{r}
# Probably easiest to import python's os library as an R object.
# as a side effect this makes sure the conda environment is working.
os <- import("os")

bash_cmd <- "which AssignGenes.py"  # AssignGenes.py is in changeo suite.

os$system(bash_cmd)
```

```{r}

# the conda_env_finalise flag is TRUE if the conda environment is built in the chunk above.
# this chunk is wrapped in an if{} so it only executes if the conda environment is new.

if(conda_env_finalise == TRUE) {

### There are two more preparatory steps
## See here: https://changeo.readthedocs.io/en/stable/examples/igblast.html

## These only need to be initiated once in the conda environment.
## they take < 1min to run, so I haven't yet stopped them running each knit.

## STEP 1 - get additional immcantation scripts ##

# 4 scripts from the Immcantation repository are required for to make igblast database from IGMT.
# 1 script is useful for clone assignment 'light_cluster.py' - used much later.
# For convenience, light_cluster.py is downloaded now.
immcant_scripts <- c("fetch_igblastdb.sh",
                     "fetch_imgtdb.sh",
                     "clean_imgtdb.py",
                     "imgt2igblast.sh",
                     "light_cluster.py")

# Make bash code to curl them all:
system(paste0("curl https://bitbucket.org/kleinstein/immcantation/raw/72aa80f6f5aeb9b211801139056e302fc8785c7b/scripts/", 
       immcant_scripts, 
       " -o ", conda_path, "/bin/", immcant_scripts,
       collapse = "\n\n")
)

# Make these executable:
system(paste0("chmod 755 ", conda_path, "/bin/", immcant_scripts,
       collapse = "\n\n"))

#### STEP 2 - make IGMT references #####

# Download reference databases
system(paste0(conda_path, "/bin/fetch_igblastdb.sh -o ", conda_path, "/share/igblast"))
system(paste0(conda_path, "/bin/fetch_imgtdb.sh -o ", conda_path, "/share/germlines/imgt"))
# Build IgBLAST database from IMGT reference sequences
system(paste0(conda_path, "/bin/imgt2igblast.sh -i ", conda_path, "/share/germlines/imgt ", "-o ", conda_path, "/share/igblast"))

}
```

```{r}
#0. make a df with filenames and annotations

file_path <- list.files(pattern = "filtered_contig_annotations.csv",
                         recursive = T, full.names = T, all.files = T)

file_path <- file_path [ grepl(file_path, pattern = "outs")]
file_path <- file_path [ grepl(file_path, pattern = "cellranger")]

labels <- data.frame(cbind(
  "SRR" = gsub(basename(file_path), pattern = ".csv", replacement =""),
  "day" = c("d60", "d60", "d28", "d28", "d12", "d12", "d12", "d5", "d5", "d0", "d0", "d60", "d28", "d12", "d5", "d0", "d60", "d28", "d12", "d5", "d0"),
  "source" = c(rep("FNA",11), rep("PBMC_enriched_IgD_neg", 5), rep("PBMC", 5))
  
)
  
)

labels$short_name <- paste(labels$source, labels$day, sep = "_")

labels$short_name <- paste0(labels$short_name,
                            ifelse(duplicated(labels$short_name), "_2", "_1"))

labels$short_name[duplicated(labels$short_name)] <- gsub(labels$short_name[duplicated(labels$short_name)], pattern = "_2", replacement = "_3")

####
# 1. make a temporary directory
####
system("mkdir temp")
setwd("temp")

####
# 2. Soft link files
####
system(
  paste0("ln -s ../cellranger/Turner_VDJ_",
       substr(labels$SRR, 1, 11),
       "/outs/",
       labels$SRR,
       ".csv ",
       labels$SRR,
       ".csv",
       collapse = "\n\n")
       )

system(
  paste0("ln -s ../cellranger/Turner_VDJ_",
       substr(labels$SRR, 1, 11),
       "/outs/filtered_contig.fasta ",
       substr(labels$SRR, 1, 11),
       "_filtered_contig.fasta",
       collapse = "\n\n")
)
####
# 3. Make a db for each SRR
####
fastas <- list.files(pattern = "filtered_contig.fasta")

lapply(fastas, function(fasta) {
os$system(
    paste0("AssignGenes.py igblast -s ",fasta,
                 " -b ",
                 conda_path, "/share/igblast/ --organism human --loci ig --format blast")
    )
}
)

#
annotations <- list.files(pattern = "filtered_contig_annotations.csv")
fmt <- list.files(pattern = ".fmt7")

os$system(
  paste0("MakeDb.py igblast -i ", fmt,
                 " -s ", fastas, 
                 " -r ", conda_path, "/share/germlines/imgt/human/vdj/",
                 " --10x ", annotations, " --extended",
         collapse ="\n\n"))


#####
# 4. Add a field to identify each one
#####
db <- list.files(pattern = "db-pass.tsv")

all(substr(db,1,11) == substr(labels$SRR, 1, 11)) # should be TRUE

os$system(
  paste0("ParseDb.py add -d ", db,
         " -f short_name -u ",
         labels$short_name, collapse = "\n\n"))
  
#####
# 5. Merge
#####
db <- list.files(pattern = "_db-pass_parse-add.tsv")

os$system(
  paste0("ParseDb.py merge -d ",
       paste0(db, collapse = " "),
" -o all_samples_merged_filtered_contig_igblast_db-pass.tsv")
)

#####
# 6. Assign clones
#####

### Split to heavy and light chains
# note that "IGH" requires " ", so the string is marked in R using ' '.
os$system(paste0('ParseDb.py select -d ',
                  "all_samples_merged_filtered_contig_igblast_db-pass.tsv",
                 ' -f locus -u "IGH" --logic all --regex --outname heavy'))

os$system(paste0('ParseDb.py select -d ',
                  "all_samples_merged_filtered_contig_igblast_db-pass.tsv",
                 ' -f locus -u "IG[LK]" --logic all --regex --outname light'))

## The changeo-10x pipeline script would now split these into {heavy, light} {productive and non-productive}.
# The changeo 10x parsing advice does not insist on spliting {prod, non-prod}
# All of the testing "filtered_contigs" contain only productive IgH, so I have not enforced this step.

### Use shazam to set clone threshold ##
# https://bitbucket.org/kleinstein/immcantation/src/72aa80f6f5aeb9b211801139056e302fc8785c7b/pipelines/shazam-threshold.R?at=master#shazam-threshold.R-2,20,22,34,50:51,69
# Above is the Rscript used to wrap the following R code.

# Calculate distance-to-nearest
library(shazam)
library(alakazam)

db <- readChangeoDb(file = "heavy_parse-select.tsv")
db <- distToNearest(db, sequenceColumn="junction", vCallColumn="v_call", jCallColumn="j_call",
                                     model="ham", first=FALSE, normalize="len")

# Calculate threshold
threshold <- findThreshold(db[["dist_nearest"]], method="density", model="gamma-gamma")
# Plot the calculated threshold:
plot(threshold, binwidth=0.02, silent=FALSE)


### Define clones ###
os$system(paste0('DefineClones.py -d heavy_parse-select.tsv --model ham --dist ', print(threshold),
        ' --mode gene --act set --outname "heavy" --format airr'))

### Light chain correction ###
os$system(paste0('light_cluster.py -d heavy_clone-pass.tsv -e light_parse-select.tsv -o 10x_clone_pass.tsv --format airr --doublets count'))
# Worth noting the --doublets count option - this takes the highest expression IgH if a cell contains >1.
# I think for filtered_contigs/_annotations cellranger has already done this filter.
# Alternative options include discarding cells with |IgH|>1.

### the 10x script would end at this point ###

### Create germlines ###
# This is required to calculate SHM, as the germline provides the reference
# It also allows us to mask the D segment - standard shazam practice.
# and it allows use to make a single germline for each clone.
##

os$system(paste0("CreateGermlines.py -d 10x_clone_pass.tsv -g dmask ",
                 "--cloned ",
                 # ^ this maps each clone back to a 'germline'
                 # we want to map each cell back to its germline, rather than rely on the clonal groups.
                 # In the earlier 'immcantation_reticulate.Rmd, which uses only IgD- PBMC @ d28, 
                 # this --cloned option was disabled (to run on a per cell basis)
       "-r ",conda_path, "/share/germlines/imgt/human/vdj/imgt_human_IGHV.fasta ",
       conda_path, "/share/germlines/imgt/human/vdj/imgt_human_IGHD.fasta ",
       conda_path, "/share/germlines/imgt/human/vdj/imgt_human_IGHJ.fasta "))





```


```{r}
# SHM mapping.
library(shazam)
library(alakazam)
db <- readChangeoDb("10x_clone_pass_germ-pass.tsv")


# Annotate mutation counts by CDRx and FRx:
db_obs <- observedMutations(db, sequenceColumn="sequence_alignment",
                            germlineColumn="germline_alignment_d_mask",
                            regionDefinition=IMGT_V_BY_REGIONS,
                            frequency=FALSE)

# Annotate mutation counts by CDR vs FR:
db_obs <- observedMutations(db_obs, sequenceColumn="sequence_alignment",
                            germlineColumn="germline_alignment_d_mask",
                            regionDefinition=IMGT_V_BY_REGIONS,
                            combine = TRUE,
                            frequency=FALSE)


save(db_obs, file = "../Full_Immcantation_db_SRRSRR11233635.RData")

```

```{r}
## Plot R and S:
## 
library(tidyverse)
db_obs %>%
  mutate("c_call" = factor(c_call)) %>%
  mutate("cdr1_ratio" = mu_count_cdr1_r / (mu_count_cdr1_s+0.01)) %>%
  mutate("cdr2_ratio" = mu_count_cdr2_r / (mu_count_cdr2_s+0.01)) %>%
  mutate("fwr1_ratio" = mu_count_fwr1_r / (mu_count_fwr1_s+0.01)) %>%
  select(cell_id, c_call, contains("ratio")) %>%
  pivot_longer(!c(cell_id, c_call), names_to = "region", values_to = "RS") %>%
  ggpubr::ggboxplot(x = "c_call", y = "RS", facet.by = "region", ylab = "R/S ratio") + ggpubr::rotate_x_text()

```

# SessionInfo

```{r sessionInfo}

sessionInfo()

```
